+++ packages/mermaid-layout-elk/src/render.ts
@@ -5,6 +5,7 @@ import type {
   SVG,
   SVGGroup,
 } from '@mermaid-chart/mermaid';
+// @ts-ignore TODO: Investigate D3 issue
 import { curveLinear } from 'd3';
 import ELK from 'elkjs/lib/elk.bundled.js';
 import { type TreeData, findCommonAncestor } from './find-common-ancestor.js';