+++ packages/mermaid/src/rendering-util/rendering-elements/shapes/anchor.ts
@@ -1,12 +1,12 @@
+import rough from 'roughjs';
 import { log } from '../../../logger.js';
-import { updateNodeBounds, getNodeClasses } from './util.js';
-import intersect from '../intersect/index.js';
-import type { Node } from '../../../rendering-util/types.d.ts';
 import {
   styles2String,
   userNodeOverrides,
-} from '../../../rendering-util/rendering-elements/shapes/handDrawnShapeStyles.js';
-import rough from 'roughjs';
+} from '../../rendering-elements/shapes/handDrawnShapeStyles.js';
+import type { Node } from '../../types.d.ts';
+import intersect from '../intersect/index.js';
+import { getNodeClasses, updateNodeBounds } from './util.js';
 
 export const anchor = (parent: SVGAElement, node: Node): Promise<SVGAElement> => {
   const { labelStyles } = styles2String(node);