+++ packages/mermaid/src/rendering-util/rendering-elements/shapes/curlyBraceRight.ts
@@ -1,11 +1,8 @@
-import { labelHelper, updateNodeBounds, getNodeClasses, createPathFromPoints } from './util.js';
-import intersect from '../intersect/index.js';
-import type { Node } from '../../../rendering-util/types.d.ts';
-import {
-  styles2String,
-  userNodeOverrides,
-} from '../../../rendering-util/rendering-elements/shapes/handDrawnShapeStyles.js';
 import rough from 'roughjs';
+import type { Node } from '../../types.d.ts';
+import intersect from '../intersect/index.js';
+import { styles2String, userNodeOverrides } from './handDrawnShapeStyles.js';
+import { createPathFromPoints, getNodeClasses, labelHelper, updateNodeBounds } from './util.js';
 
 function generateCirclePoints(
   centerX: number,