+++ packages/mermaid/src/rendering-util/rendering-elements/shapes/filledCircle.ts
@@ -1,11 +1,11 @@
-import { log } from '../../../logger.js';
-import { getNodeClasses, updateNodeBounds } from './util.js';
-import type { Node } from '../../../rendering-util/types.d.ts';
-import type { SVG } from '../../../diagram-api/types.js';
-import { userNodeOverrides } from '../../../rendering-util/rendering-elements/shapes/handDrawnShapeStyles.js';
 import rough from 'roughjs';
-import intersect from '../intersect/index.js';
 import { getConfig } from '../../../config.js';
+import type { SVG } from '../../../diagram-api/types.js';
+import { log } from '../../../logger.js';
+import type { Node } from '../../types.d.ts';
+import intersect from '../intersect/index.js';
+import { userNodeOverrides } from './handDrawnShapeStyles.js';
+import { getNodeClasses, updateNodeBounds } from './util.js';
 
 export const filledCircle = (parent: SVG, node: Node) => {
   node.label = '';