+++ packages/mermaid/src/rendering-util/rendering-elements/shapes/forkJoin.ts
@@ -1,6 +1,6 @@
 import rough from 'roughjs';
 import type { SVG } from '../../../diagram-api/types.js';
-import type { Node } from '../../types.js';
+import type { Node } from '../../types.d.ts';
 import intersect from '../intersect/index.js';
 import { styles2String, userNodeOverrides } from './handDrawnShapeStyles.js';
 import { getNodeClasses, updateNodeBounds } from './util.js';
@@ -51,6 +51,5 @@ export const forkJoin = (parent: SVG, node: Node, dir: string) => {
   node.intersect = function (point) {
     return intersect.rect(node, point);
   };
-
   return shapeSvg;
 };