+++ packages/mermaid/src/rendering-util/rendering-elements/shapes/lightningBolt.ts
@@ -1,10 +1,10 @@
-import { log } from '../../../logger.js';
-import { getNodeClasses, updateNodeBounds, createPathFromPoints } from './util.js';
-import type { Node } from '../../../rendering-util/types.d.ts';
-import type { SVG } from '../../../diagram-api/types.js';
-import { userNodeOverrides } from '../../../rendering-util/rendering-elements/shapes/handDrawnShapeStyles.js';
 import rough from 'roughjs';
+import type { SVG } from '../../../diagram-api/types.js';
+import { log } from '../../../logger.js';
+import type { Node } from '../../types.d.ts';
 import intersect from '../intersect/index.js';
+import { userNodeOverrides } from './handDrawnShapeStyles.js';
+import { createPathFromPoints, getNodeClasses, updateNodeBounds } from './util.js';
 
 export const lightningBolt = (parent: SVG, node: Node) => {
   node.label = '';